requirements.txt
asgiref==3.9.1
certifi==2025.8.3
cffi==2.0.0
charset-normalizer==3.4.3
cryptography==45.0.7
Django==5.2.6
django-cors-headers==4.8.0
django-oauth-toolkit==3.0.1
djangorestframework==3.16.1
drf-yasg==1.21.10
idna==3.10
inflection==0.5.1
jwcrypto==1.5.6
oauthlib==3.3.1
packaging==25.0
psycopg2==2.9.10
psycopg2-binary==2.9.10
pycparser==2.23
pytz==2025.2
PyYAML==6.0.2
requests==2.32.5
sqlparse==0.5.3
typing_extensions==4.15.0
uritemplate==4.2.0
urllib3==2.5.0
